Planx-eSearch/py-planx-eSearch-data-feed/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptg==0.5.0.post0
dataclasses-json==0.6.7
datasets==3.3.2
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.17.0
Flask==3.1.0
frozenlist==1.5.0
fsspec==2024.12.0
googlesearch-python==1.3.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
kafka-python==2.0.5
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
langsmith==0.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.63.2
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.2
panda==0.3.1
pandas==2.2.3
playwright==1.50.0
propcache==0.3.0
pyaes==1.6.1
pyarrow==19.0.1
pyasn1==0.6.1
pybase62==1.0.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyee==12.1.1
pymongo==4.11.2
pyperclip==1.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-socks==2.7.1
pytz==2025.1
PyYAML==6.0.2
ragas==0.2.13
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.46.0
Telethon==1.39.0
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
unicorn==2.1.2
urllib3==2.3.0
uvicorn==0.34.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
